backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-rest-knox==4.2.0
dj-database-url==3.1.2
Pillow==10.4.0
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.11.0
django-filter==23.5
cryptography==42.0.5
pytz==2024.1